+++ package.json
@@ -28,6 +28,7 @@
     "lucide-react": "^0.364.0",
     "mini-svg-data-uri": "^1.4.4",
     "next": "14.1.4",
+    "next-intl": "^4.3.4",
     "next-themes": "^0.3.0",
     "react": "^18",
     "react-dom": "^18",
