+++ src/app/[locale]/about/AboutWrapper.tsx
