+++ src/app/not-found.tsx
@@ -1,0 +1,19 @@
+import { NextIntlClientProvider } from 'next-intl';
+import { getMessages } from 'next-intl/server';
+
+export default async function NotFound() {
+  const messages = await getMessages();
+
+  return (
+    <html lang="en">
+      <body>
+        <NextIntlClientProvider messages={messages}>
+          <div className="flex flex-col items-center justify-center min-h-screen">
+            <h1 className="text-4xl font-bold mb-4">404 - Page Not Found</h1>
+            <p className="text-lg text-gray-600">The page you're looking for doesn't exist.</p>
+          </div>
+        </NextIntlClientProvider>
+      </body>
+    </html>
+  );
+}
