+++ src/components/ui/timeline.tsx
@@ -6,6 +6,7 @@ import {
   motion,
 } from "framer-motion";
 import React, { useEffect, useRef, useState } from "react";
+import { useTranslations } from 'next-intl';
 
 interface TimelineEntry {
   title: string;
@@ -13,6 +14,7 @@ interface TimelineEntry {
 }
 
 export const Timeline = ({ data }: { data: TimelineEntry[] }) => {
+  const t = useTranslations('AboutPage');
   const ref = useRef<HTMLDivElement>(null);
   const containerRef = useRef<HTMLDivElement>(null);
   const [height, setHeight] = useState(0);
@@ -39,11 +41,10 @@ export const Timeline = ({ data }: { data: TimelineEntry[] }) => {
     >
       <div className="max-w-7xl mx-auto py-40 px-4 md:px-8 lg:px-10">
         <h2 className="text-lg md:text-4xl mb-4 text-black dark:text-white max-w-4xl">
-          Timeline of BuzzVerse
+          {t('timelineTitle')}
         </h2>
         <p className="text-neutral-700 dark:text-neutral-300 text-sm md:text-base max-w-sm">
-          We&apos;ve been working on BuzzVerse for the past 3 years. Here&apos;s
-          a timeline of our science club.
+          {t('timelineDescription')}
         </p>
       </div>
